update package-lock.json 2
